Moved click_handler
