+++ lib/src/logic/interaction/click_handler.dart
